requirements.txt
streamlit
pandas
matplotlib
requests
python-dotenv
rapidfuzz